changed the name of index.test.ts to todoApi.test.ts and created a describe block for the todoAPI test
